requirements.txt
django
djangorestframework
pillow
python-jose
djangorestframework-jwt
python-dotenv
gunicorn
boto3
django-storages
requests